+++ src/stringView.c
@@ -94,7 +94,7 @@ int64_t svFirstIndexOf(StringView src, StringView str) {
 
 int svCmp(StringView a, StringView b) {
     if (a.size != b.size)
-        return a.size - b.size;
+        return (int)(a.size - b.size);
     // NOTE: memcmp works exactly the same here but doesn't stop at null
     // characters. In regular use there is no difference between the two
     return memcmp(a.data, b.data, a.size);
